missing files added and order model edited

+++ NodeJS Application/models/Cart.js
@@ -14,7 +14,7 @@ const Cart = sequelize.define("cart", {
         defaultValue: 0,
     },
     totalPrice: {
-        type: Sequelize.DECIMAL(20, 2),
+        type: Sequelize.INTEGER,
         defaultValue: 0,
         allowNull: false,
         trim: true,
